+++ example/lib/main.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:image_editor_plus/image_editor_plus.dart';
-import 'package:image_editor_plus/utils.dart';
 
 void main() {
   runApp(
@@ -76,18 +75,6 @@ class _ImageEditorExampleState extends State<ImageEditorExample> {
                       imageData,
                       imageData,
                     ],
-                    features: const ImageEditorFeatures(
-                      pickFromGallery: true,
-                      captureFromCamera: true,
-                      crop: true,
-                      blur: true,
-                      brush: true,
-                      emoji: true,
-                      filters: true,
-                      flip: true,
-                      rotate: true,
-                      text: true,
-                    ),
                   ),
                 ),
               );